external-access-static-host-based updates
ingress-service-hostbased.yaml: networking.k8s.io/v1beta1 is deprecated Kubernetes v1.22. Currently v1 is only supported

README.rst: The Helm release is called "ingress-nginx"

topic.yaml: In this example "min.insync.replicas" is override to "2". The option "replicas: 1" makes the topic to be unavailable so the producer gets an error when trying to write.