Update - external-access-static-host-based (#169)
* external-access-static-host-based updates

ingress-service-hostbased.yaml: networking.k8s.io/v1beta1 is deprecated Kubernetes v1.22. Currently v1 is only supported

README.rst: The Helm release is called "ingress-nginx"

topic.yaml: In this example "min.insync.replicas" is override to "2". The option "replicas: 1" makes the topic to be unavailable so the producer gets an error when trying to write.

* Update - external-access-static-host-based

ingress-service-hostbased.yaml: networking.k8s.io/v1beta1 is deprecated in Kubernetes v1.22. Currently, networking.k8s.io/v1 is only supported for the new Kubernetes versions.

README.rst: The Helm release is called "ingress-nginx" instead of "nginx-operator"

topic.yaml: In this example, the brokers are configured with "min.insync.replicas = 2". I've observed that the option "replicas: 1" makes the topic to be unavailable so the producer gets an error when trying to write to topic "elastic-0". Removing that option the issue is solved.